# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.FacilityCreateDTO:
    properties:
      address:
        type: string
      contact_email:
        type: string
      contact_person:
        type: string
      contact_phone:
        type: string
      dhis2_code:
        type: string
      district:
        type: string
      emr_system_code:
        type: string
      emr_system_name:
        type: string
      facility_code:
        type: string
      facility_name:
        type: string
      is_active:
        type: boolean
      level_of_care:
        type: string
      region:
        type: string
      zone:
        type: string
    required:
    - facility_code
    - facility_name
    type: object
...
  dto.FacilityResponseDTO:
    properties:
      address:
        type: string
      contact_email:
        type: string
      contact_person:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      dhis2_code:
        type: string
      district:
        type: string
      emr_system_code:
        type: string
      emr_system_name:
        type: string
      facility_code:
        type: string
      facility_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      level_of_care:
        type: string
      pharmacies:
        items:
          $ref: '#/definitions/dto.PharmacyResponseDTO'
        type: array
      region:
        type: string
      updated_at:
        type: string
      zone:
        type: string
    type: object
  dto.FacilityUpdateDTO:
    properties:
      address:
        type: string
      contact_email:
        type: string
      contact_person:
        type: string
      contact_phone:
        type: string
      dhis2_code:
        type: string
      district:
        type: string
      emr_system_code:
        type: string
      emr_system_name:
        type: string
      facility_name:
        type: string
      is_active:
        type: boolean
      level_of_care:
        type: string
      region:
        type: string
      zone:
        type: string
    type: object
  dto.GoodsReceiptCreateDTO:
    properties:
      grn_batch_number:
        type: string
      grn_expiry_date:
        type: string
      grn_facility_code:
        type: string
      grn_facility_receipt_number:
        type: string
      grn_order_number:
        type: string
      grn_product_code:
        type: string
      grn_quantity:
        type: integer
      grn_receipt_date:
        type: string
      grn_supplier_code:
        type: string
      grn_system_code:
        type: string
      grn_warehouse_ref_number:
        type: string
    required:
    - grn_batch_number
    - grn_expiry_date
    - grn_facility_code
    - grn_facility_receipt_number
    - grn_order_number
    - grn_product_code
    - grn_quantity
    - grn_receipt_date
    - grn_supplier_code
    - grn_system_code
    - grn_warehouse_ref_number
    type: object
  dto.GoodsReceiptResponseDTO:
    properties:
      grn_batch_number:
        type: string
      grn_expiry_date:
        type: string
      grn_facility_code:
        type: string
      grn_facility_receipt_number:
        type: string
      grn_order_number:
        type: string
      grn_product_code:
        type: string
      grn_quantity:
        type: integer
      grn_receipt_date:
        type: string
      grn_supplier_code:
        type: string
      grn_system_code:
        type: string
      grn_timestamp:
        type: string
      grn_warehouse_ref_number:
        type: string
      id:
        type: integer
      sync_status:
        type: integer
      validation_status:
        type: integer
    type: object
  dto.GoodsReceiptUpdateDTO:
    properties:
      grn_quantity:
        type: integer
    type: object
...
  dto.PharmacyResponseDTO:
    properties:
      created_at:
        type: string
      facility:
        $ref: '#/definitions/dto.FacilityResponseDTO'
      facility_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      pharmacy_code:
        type: string
      pharmacy_name:
        type: string
      pharmacy_type:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: ict@moh.go.ug
    name: DHI
  description: Backend API for Supply Chain Repository / Bridge
  termsOfService: https://health.go.ug
  title: MoH Supply Chian Repository API
  version: "1.0"
paths:
...
  /facilities:
    get:
      parameters:
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Filter by region
        in: query
        name: region
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.FacilityResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List facilities
      tags:
      - Facility
    post:
      consumes:
      - application/json
      parameters:
      - description: Facilities payload
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.FacilityCreateDTO'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.FacilityResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create facilities
      tags:
      - Facility
  /facilities/{id}:
    delete:
      parameters:
      - description: Facility ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete facility
      tags:
      - Facility
    get:
      parameters:
      - description: Facility ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FacilityResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get facility by ID
      tags:
      - Facility
    put:
      consumes:
      - application/json
      parameters:
      - description: Facility ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.FacilityUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FacilityResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update facility
      tags:
      - Facility
...
  /goods-receipt:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GoodsReceiptResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List goods receipts
      tags:
      - GoodsReceipt
    post:
      consumes:
      - application/json
      parameters:
      - description: Goods receipts payload
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.GoodsReceiptCreateDTO'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.GoodsReceiptResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create goods receipts
      tags:
      - GoodsReceipt
  /goods-receipt/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete goods receipt
      tags:
      - GoodsReceipt
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GoodsReceiptResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get goods receipt by id
      tags:
      - GoodsReceipt
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.GoodsReceiptUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GoodsReceiptResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update goods receipt
      tags:
      - GoodsReceipt
...
swagger: "2.0"